move connection utils
