+++ cmd/create.go
@@ -88,7 +88,7 @@ var createServiceCmd = &cobra.Command{
 		if err != nil {
 			return err
 		}
-		unAuthConn, err := newConnection()
+		unAuthConn, err := utils.NewAnonymousConnection()
 		if err != nil {
 			return err
 		}
@@ -99,7 +99,7 @@ var createServiceCmd = &cobra.Command{
 			return err
 		}
 
-		conn, err := newAuthenticatedConnection(accessToken)
+		conn, err := utils.NewAuthenticatedConnection(accessToken)
 		if err != nil {
 			return err
 		}
